Move utility functions to lokomotive package.
- As part of the refactor, all Lokomotive cluster related utility
  functions are ported to lokomotive package.

  Lokomotive package holds all the code related to management of
  Lokomtive cluster.
